##Introduction
#this is the Elite Techno Groups Internship Assignment Here we used Summer olympics dataset for Analysis and solving the following questions

In how many cities Summer Olympics is held so far?
Which sport is having most number of Gold Medals so far?
Which sport is having most number of medals so far? 
Which player has won most number of medals? 
Which player has won most number Gold Medals of medals?
In which year India won first Gold Medal in Summer Olympics?
Which event is most popular in terms on number of players? 
Which sport is having most female Gold Medalists?